Remove bilayer functionality and refactor pipeline
- Removed create_bilayer_perovskite from pipeline and creator
- Refactored common pipeline logic into pipeline/common.py
- Renamed population.py to populate.py for consistency
- Added support for 'HOLE' reserved word in spacer lists
- Updated documentation for spacer=None behavior
